pyproject.toml
[tool.poetry]
name = "noteoverflow"
version = "0.1.0"
description = ""
authors = ["BiDuang <me@biduang.cn>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.2"
uvicorn = "^0.23.2"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
sqlalchemy = "^2.0.21"
python-multipart = "^0.0.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
